Extract subagent rollup logic from state.py
